Update Jet PF ID
Following the recommandation here : https://twiki.cern.ch/twiki/bin/view/CMS/JetID13TeVRun2016 (27-dec-2016) I've updated the jetID.

For the 13 TeV analysis for promt and rereco data of Run2016 the same recommendations as before are valid for |eta|<2.7 and 3<|eta|<5.

For 2.7< |eta|<3 the JetID criteria have been changed (see commit) in order to increase the efficiency and background rejection as well. For more details look at the twiki.
For |eta|>2.7 Loose and Tight criteria are identical as before.